Rename project to 'scrubbish'
It was going to be 'scrapy', but sure that's already
taken by a more vererable project. Then I thought to
use 'skrapy' to get over this fact.

Finally I thought, "sure this is a crappy, rubbish
standard of coding and it's kind of doing my head
in... 'scrubbish' will do".